+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Broker's configuration format changed from INI to TOML
+
+
 ## [0.6.1] - 2024-04-12
 ### Added
 - `is_shvtype` and `is_shvlist` functions that validates Python object as being